pyproject.toml
[tool.poetry]
name = "testing-python"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"github3.py" = "^1.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
